update user info,avatar

+++ src/main/java/com/javaweb/utils/Constant.java
@@ -1,6 +1,6 @@
 package com.javaweb.utils;
 
 public class Constant {
-    public static final long STAFF_ROLE = 0;
-    public static final long ADMIN_ROLE = 1;
+    public static final int STAFF_ROLE = 0;
+    public static final int ADMIN_ROLE = 1;
 }
